// frontend/src/generated/releaseNotes.ts
export type ReleaseChange = {
  type: string
  label: string
  scope: string | null
  breaking: boolean
  subject: string
  pr: number | null
}

export type Release = {
  sha: string
  date: string
  pr: number | null
  changes: ReleaseChange[]
}